+++ src/core/trezor/initializeTrezor.ts
@@ -1,0 +1,12 @@
+export const initializeTrezor = () => {
+  chrome.runtime.onInstalled.addListener(() => {
+    TrezorConnect.init({
+      manifest: {
+        email: 'support@rainbow.me',
+        appUrl: 'https://rainbow.me',
+      },
+      transports: ['BridgeTransport', 'WebUsbTransport'],
+      connectSrc: 'https://connect.trezor.io/9/',
+    });
+  });
+};
