+++ src/entries/popup/index.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <title>Rainbow Wallet</title>
-  <script src="./vendor/trezor-connect.js" type="text/javascript"></script>
+  <script src="./vendor/trezor-connect-webextension.js" type="text/javascript"></script>
 </head>
 
 <body>
